UI-25 export types
