+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "cohesive-ui",
-  "version": "1.0.0",
+  "version": "1.0.2",
   "main": "dist/index.js",
   "files": [
     "dist"
@@ -45,7 +45,6 @@
   },
   "scripts": {
     "build": "yarn webpack --mode=production --types",
-    "prepare": "yarn test && yarn build && cp -r src/style dist/",
     "build-storybook": "build-storybook -c .storybook -o docs",
     "test": "jest",
     "storybook": "start-storybook -p 6006"
